Move out common JS modules for assessment form
