+++ caseworker/assets/javascripts/assessment-form/__tests__/select-products.test.js
