+++ caseworker/assets/javascripts/assessment-form/select-products.js
