+++ caseworker/assets/javascripts/tau.js
@@ -2,8 +2,8 @@ import SelectAll, { SELECT_ALL_BUTTON_TEXT } from "core/select-all";
 import ExpandAll, { SHOW_ALL_BUTTON_TEXT } from "core/expand-all";
 import CheckboxClassToggler from "core/checkbox-class-toggler";
 import DisablingButton from "core/disabling-button";
-import Headline from "./tau/headline";
-import SelectProducts from "./tau/select-products";
+import Headline from "./assessment-form/headline";
+import SelectProducts from "./assessment-form/select-products";
 import CLESuggestions from "./tau/cle-suggestions";
 import SuggestionsTokenField from "./tau/suggestions-token-field";
 import NoSuggestionsTokenField from "./tau/no-suggestions-token-field";
